move SubmissionList.vue to components/submission move ProblemCode.vue to components/problem move ContestProcess.vue to components/contest
