+++ src/components/submission/SubmissionList.vue
@@ -55,7 +55,7 @@
 </template>
 
 <script>
-import ProblemCode from '_c/ProblemCode';
+import ProblemCode from '_c/problem/ProblemCode';
 import JudgeResult from '_c/JudgeResult';
 import { Websocket } from '_c/mixins';
 import { JUDGE_RESULT_TYPE, CONTEST_STATUS } from '_u/constants';
