+++ src/views/contest/ContestDetailView.vue
@@ -111,7 +111,7 @@ import { mapGetters, mapState } from 'vuex';
 import { CONTEST_OPENNESS, CONTEST_STATUS } from '_u/constants';
 import { s2hs } from '_u/transform';
 
-import ContestProcess from '_c/ContestProcess';
+import ContestProcess from '_c/contest/ContestProcess';
 import Markdown from '_c/editor/Markdown';
 
 import rankHandler from '@/store/modules/ranks';
